+++ src/main/java/subway/line/LineController.java
@@ -4,6 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.section.Section;
+import subway.section.SectionDao;
+import subway.section.SectionRequest;
+import subway.section.Sections;
 import subway.station.StationDao;
 import subway.station.Stations;
