+++ src/main/java/subway/section/SectionDao.java
@@ -1,8 +1,9 @@
-package subway.line;
+package subway.section;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
+import subway.line.LineRequest;
 
 @Repository
 public class SectionDao {
