+++ src/main/java/subway/section/Sections.java
@@ -1,6 +1,5 @@
-package subway.line;
+package subway.section;
 
-import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
