+++ src/test/java/subway/line/SectionAcceptanceTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
+import subway.section.SectionRequest;
 import subway.station.StationResponse;
 
 import java.util.Arrays;
